cli/pyproject.toml
[tool.poetry]
name = "tin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]
include = ["pyproject.toml"]

[tool.poetry.scripts]
tin = "app.main:app"
api = "uvicorn:run(app='app.api.main:app', host='0.0.0.0', port=5000, reload=True)"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-multipart = "^0.0.18"
openai = "^1.12.0"
pandas = "^2.2.0"
python-dotenv = "^1.0.0"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
